+++ chb/ast/ASTCPrettyPrinter.py
@@ -386,6 +386,9 @@ class ASTCPrettyPrinter(ASTVisitor):
         self.ccode.newline(indent=self.indent)
         self.ccode.write("default:")
 
+    def visit_nop_instr(self, instr: AST.ASTNOPInstruction) -> None:
+        self.ccode.write("// " + instr.description)
+
     def visit_assign_instr(self, instr: AST.ASTAssign) -> None:
         self.ccode.newline(indent=self.indent)
         instr.lhs.accept(self)