+++ chb/ast/ASTTransformer.py
@@ -105,6 +105,11 @@ class ASTTransformer(ABC):
         ...
 
     @abstractmethod
+    def transform_nop_instr(
+            self, instr: AST.ASTNOPInstruction) -> AST.ASTInstruction:
+        ...
+
+    @abstractmethod
     def transform_lval(self, lval: AST.ASTLval) -> AST.ASTLval:
         ...