+++ chb/ast/doc/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Version 0.1.0-2023-05-10
+
+- Add ASTNOPInstruction as a subclass of ASTInstruction to enable connecting ast
+  elements without semantic content with span information
+
+---
 
 ### Version 0.1.0-2023-05-02